Modified: 1) AllDifferentElementConstraint::calculateSolutionScore previous version had a bug w/ score: solutions with only 2 items scored too much (if 6 nodes in the graph and 2 items repeating, score was 4 (D:) now it scores 0 (zero), because items have not to repeat on solutions) 2) Genetic Solver added mutation (one-point mutation), elitism (that needs sorting of the whole vector of population: this may make computation really slow) and some microimprovements on population management (in main function) 3) maybe something else I forgot to say there :O

+++ RoadRage/solvers/genetic/geneticengine.cpp
@@ -9,6 +9,9 @@
 GeneticEngine::GeneticEngine(unsigned int generations)
 {
     this->generations=generations;
+    this->elitistsNumber=0;
+    this->crossover=0;
+    this->gih=NULL;
 }
 
 void
@@ -67,6 +70,17 @@ GeneticEngine::getNewIndividualsNumber()
     return this->newIndividualsNumber;
 }
 
+void
+GeneticEngine::setElitismNumber(unsigned int elitists)
+{
+    this->elitistsNumber=elitists;
+}
+unsigned int
+GeneticEngine::getElitismNumber()
+{
+    return this->elitistsNumber;
+}
+
 Path
 GeneticEngine::generateRandomPath(unsigned int solutionLength)
 {
@@ -114,7 +128,7 @@ GeneticEngine::evaluatePopulation(QVector<Path>* population,
     {
         Path candidate=population->at(popIndex);
         individualScore=0;
-        unsigned int newValue;
+        //unsigned int newValue;
         foreach(Constraint* constr, constraints)
         {
 
@@ -127,9 +141,42 @@ GeneticEngine::evaluatePopulation(QVector<Path>* population,
 }
 
 void
+GeneticEngine::mutatePopulation(QVector<Path> *population)
+{
+    srand(3);
+    for(unsigned int p_i=0; p_i<(population->size()-this->elitistsNumber); p_i++)
+    {
+        if((rand()/RAND_MAX)<this->mutationProbability)
+        {
+            //workaround. why have I decided to use objects instead of pointers?
+            Path oldPath=population->at(p_i);
+            Path newPath(oldPath.getNodes(), oldPath.getPath());
+            for(int i=0; i<oldPath.getNodes(); i++)
+                newPath.addName(oldPath.getNameByPoint(i));
+            //one-point mutation
+            newPath.replace(
+                        rand()%newPath.getNodes(),
+                        rand()%newPath.getNodes()
+                        );
+            //random mutation
+            //for(int i=0; i<newPath.getNodes(); i++)
+            //{
+                //if(rand()/RAND_MAX<this->mutationProbability)
+                //    newPath.replace(rand()%newPath.getNodes()
+                //                    ,rand()%newPath.getNodes());
+            //}
+            population->replace(p_i, newPath);
+        }
+    }
+    //cout;
+}
+
+void
 GeneticEngine::sortPopulation(QVector<Path>* population, QVector<unsigned int>* scoreOfPopulation,
                               unsigned int itemsNo, bool sortBetter)
 {
+    /*
+      //from now (15-5-2012 1532) this comment is outdated.
     //it's not a real sort, just searching best(or worst) m individuals
     //and we don't care about others, so we don't have to spend O(n**2)
     //when we can use O(n*m) AND n>>m.
@@ -139,10 +186,11 @@ GeneticEngine::sortPopulation(QVector<Path>* population, QVector<unsigned int>* 
     //first time we select the 1st and put it into 1st place
     //second time we select the 1st best of the remaining and put it at 2nd place
     //and so on...
-
+    */
     int bestPosition[omp_get_max_threads()];
 
-    for(int pos=0; pos<this->crossover+this->newIndividualsNumber; pos++)
+    //for(int pos=0; pos<this->crossover+this->newIndividualsNumber; pos++)
+    for(int pos=0; pos<population->size(); pos++)
     {
 
         for(int threadIndex=0; threadIndex<omp_get_max_threads(); threadIndex++)
@@ -199,6 +247,8 @@ GeneticEngine::generateChildren(QVector<Path>* population, QVector<Path>* childr
     srand(0);
     for(unsigned int childNo=0; childNo<this->crossover/2; childNo++)
     {
+        //RAND_MAX division needed!
+        //#pop*rand()/RAND_MAX --> #pop*(a number between 0 and 1) --> result<#pop (or at least equal to)
         Path firstParent=population->at((population->size()*rand())/RAND_MAX);
         Path secondParent=population->at((population->size()*rand())/RAND_MAX);
         unsigned int crossoverPoint=rand()%this->gih->getNodes().size();
@@ -214,15 +264,11 @@ GeneticEngine::generateChildren(QVector<Path>* population, QVector<Path>* childr
             firstChild.appendPoint(item<crossoverPoint?
                                        firstParent.getPath().at(item):
                                        secondParent.getPath().at(item));
-            if(rand()/RAND_MAX<this->mutationProbability)
-                firstChild.getPath().replace(item, rand()%this->gih->getNodes().size());
-
             secondChild.appendPoint(item<crossoverPoint?
                                        secondParent.getPath().at(item):
                                        firstParent.getPath().at(item));
-            if(rand()/RAND_MAX<this->mutationProbability)
-                secondChild.getPath().replace(item, rand()%this->gih->getNodes().size());
         }
+
         //cout<<"DEBUG into generateChildren, printing children"<<endl;
         //firstChild.print();
         //secondChild.print();
@@ -249,13 +295,16 @@ GeneticEngine::getBestPaths()
     {
         start=omp_get_wtime();
         cout<<"generation #"<<times<<endl;
-        QVector<Path> *children=new QVector<Path>;
-        //people are mating.
-        cout<<"people are mating."<<endl;
-        double start_gen=omp_get_wtime();
-        this->generateChildren(population, children);
-        double end_gen=omp_get_wtime();
-        cout<<"tempo per la generazione dei figli: "<<end_gen-start_gen<<endl;
+
+
+        if(times)
+        {
+            cout<<"mutation of some individuals"<<endl;
+
+            this->mutatePopulation(population);
+            cout<<"end of the mutation"<<endl;
+        }
+
         //all can die. who's next? Let's see...
         cout<<"evaluation."<<endl;
         double start_eval=omp_get_wtime();
@@ -263,10 +312,10 @@ GeneticEngine::getBestPaths()
         {
             this->evaluatePopulation(population, scoreOfPopulation,
                                  this->populationDimension);
-            for(int i=0; i<this->populationDimension; i++)
-            {
-                Path p=population->at(i); p.print();
-            }
+            //for(unsigned int i=0; i<this->populationDimension; i++)
+            //{
+            //    Path p=population->at(i); p.print();
+            //}
         }
         else
             //we have already evaluated survived individuals
@@ -287,23 +336,35 @@ GeneticEngine::getBestPaths()
         double end_sort=omp_get_wtime();
         cout<<"tempo per il sort -parola grossa D:-: "<<end_sort-start_sort<<endl;
         //... and kill who is not strong enough to survive.
-        cout<<"kill people not strong enough to survive!"<<endl;
-        this->killIndividualsWithLowFitness(population);
-        //now replacing died with fresh meat.
-        cout<<"replacing died with children"<<endl;
-        for(unsigned int childIndex=0; childIndex<(unsigned)children->size(); childIndex++)
+        if(times+1!=this->generations)
         {
-            population->push_front(children->at(childIndex));
+
+            QVector<Path> *children=new QVector<Path>;
+            //people are mating.
+            cout<<"people are mating."<<endl;
+            double start_gen=omp_get_wtime();
+            this->generateChildren(population, children);
+            double end_gen=omp_get_wtime();
+            cout<<"tempo per la generazione dei figli: "<<end_gen-start_gen<<endl;
+            cout<<"kill people not strong enough to survive!"<<endl;
+            this->killIndividualsWithLowFitness(population);
+            //now replacing died with fresh meat.
+            cout<<"replacing died with children"<<endl;
+            for(unsigned int childIndex=0; childIndex<(unsigned)children->size(); childIndex++)
+            {
+                population->push_front(children->at(childIndex));
+            }
+            cout<<"foreign people are coming to town. adding them to population!"<<endl;
+            for(unsigned int foreignIndex=0; foreignIndex<this->newIndividualsNumber; foreignIndex++)
+                population->push_front(this->generateRandomPath(this->gih->getNodes().size()));
         }
-        cout<<"foreign people are coming to town. adding them to population!"<<endl;
-        for(unsigned int foreignIndex=0; foreignIndex<this->newIndividualsNumber; foreignIndex++)
-            population->push_front(this->generateRandomPath(this->gih->getNodes().size()));
         end=omp_get_wtime();
         cout<<end-start<<"seconds were spent on the calculation of this generation"<<endl;
         average=(average+(end-start))/2.0;
     }
     //now sort just best solutions and delete the others.
-    cout<<"average time for generation: "<<average<<endl;
+    cout<<"average time for generation: "<<average<<", generations: ";
+    cout<<this->generations<<endl;
     return *population;
 }
 
@@ -316,13 +377,13 @@ GeneticEngine::populationStats(QVector<Path> *population, QVector<unsigned int> 
     //QVector<unsigned int> scores(positionsNo);
     unsigned int *scores=(unsigned int*)malloc(positionsNo* sizeof(unsigned int));
 #pragma omp parallel for
-    for(unsigned int clear_index=0; clear_index<population->size(); clear_index++)
+    for(int clear_index=0; clear_index<population->size(); clear_index++)
     {
         scores[clear_index]=0;
     }
     //scores.fill(0);
 #pragma omp parallel for
-    for(unsigned int i=0; i<scoreOfPopulation->size(); i++)
+    for(int i=0; i<scoreOfPopulation->size(); i++)
     {
         //cout<<"item #"<<i<<endl;
         //Path lolpath=population->at(i);
@@ -340,4 +401,32 @@ GeneticEngine::populationStats(QVector<Path> *population, QVector<unsigned int> 
         if(scores[index]!=0)
             cout<<"there are #"<<scores[index]<<" items that has score "<<index<<endl;
     }
+}
+
+
+int
+GeneticEngine::compareSolutions(Path first, Path last)
+{
+    /*
+      it lets you avoid writing boilerplate code
+      in order to compare solutions using GeneticEngine::evaluatePopulation
+
+      it has a behaviour like Java's Object.compareTo(Object, Object):
+      -1 if first's score < last's score
+      0 if first's score == last's score
+      1 if first's score > last's score
+    */
+
+    QVector<Path> solutionsVector(2);
+    solutionsVector.append(first); solutionsVector.append(last);
+    QVector<unsigned int> scores(2);
+    scores.fill(0);
+
+    this->evaluatePopulation(&solutionsVector, &scores, 2);
+
+    if(scores.at(0)>scores.at(1))
+        return 1;
+    else if(scores.at(0)==scores.at(1))
+        return 0;
+    return -1;
 }
